Implement matrix multiplication
AAAAAAAAAAAAAAAAAAAAAAAAAAAAA. So it was at this point that I realized
that trying to use a straight array was NOT worth it. So I went to two-
dimensional arrays so that the math was less maddening.

I might switch back if I can work everything out cleanly. I'm a fool
like that.