Rename matrix type to 'matrix4'
This is a "just in case" change for when I need to deal with other
matrix types.

Just in case... Can you tell I'm starting to lose it from all this yak
shaving nonsense? I wouldn't be putting up with this if somebody had
told me I should have been using C++.